Working out api

+++ src/main/java/com/rumpus/common/IModel.java
@@ -8,7 +8,7 @@ public interface IModel<T extends IRumpusObject> extends IRumpusObject {
     T create();
     boolean init(Map<String, String> initList);
     Supplier<T> createFunction();
-    int getId();
-    void setId(int id);
+    Long getId();
+    void setId(Long id);
     void map(ResultSet rs);
 }
