+++ src/test/java/com/rumpus/rumpus/models/UserTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class UserTest {
-    int id = 1;
+    Long id = Long.valueOf(1);
     User user = User.createUser(id, "Frodo");
 
     @BeforeAll
